+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/TigerDirector.java
@@ -37,13 +37,12 @@ import de.gematik.test.tiger.testenvmgr.TigerTestEnvMgr;
 import de.gematik.test.tiger.testenvmgr.TigerTestEnvMgrApplication;
 import de.gematik.test.tiger.testenvmgr.controller.TestExecutionController;
 import de.gematik.test.tiger.testenvmgr.data.BannerType;
-import de.gematik.test.tiger.testenvmgr.env.ScenarioReplayer;
+import de.gematik.test.tiger.testenvmgr.env.ScenarioRunner;
 import de.gematik.test.tiger.testenvmgr.env.TigerStatusUpdate;
 import de.gematik.test.tiger.testenvmgr.servers.log.TigerServerLogManager;
 import de.gematik.test.tiger.testenvmgr.util.TigerEnvironmentStartupException;
 import de.gematik.test.tiger.testenvmgr.util.TigerTestEnvException;
 import io.cucumber.core.plugin.report.SerenityReporterCallbacks;
-import io.cucumber.core.runtime.Runtime;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
@@ -88,7 +87,6 @@ public class TigerDirector {
 
   @Getter private static TigerLibConfig libConfig;
   private static ConfigurableApplicationContext envMgrApplicationContext;
-  private static Runtime runtime;
 
   public static synchronized void start() {
     if (initialized) {
@@ -110,7 +108,6 @@ public class TigerDirector {
     try {
       // get free port
       startTestEnvMgr();
-      setupScenarioReplayer(runtime);
       startWorkflowUi();
       setupTestEnvironment(Optional.empty());
       setDefaultProxyToLocalTigerProxy();
@@ -128,11 +125,6 @@ public class TigerDirector {
     }
   }
 
-  private static void setupScenarioReplayer(Runtime runtime) {
-    ScenarioReplayer scenarioReplayer = envMgrApplicationContext.getBean(ScenarioReplayer.class);
-    scenarioReplayer.setRuntime(runtime);
-  }
-
   public static synchronized void startStandaloneTestEnvironment() {
     log.info("Starting Tiger testenvironment in STANDALONE MODE!");
     if (initialized) {
@@ -144,7 +136,7 @@ public class TigerDirector {
       readConfiguration();
       if (getLibConfig().isActivateWorkflowUi()) {
         log.warn(
-            "Starting WorkflowUI in standalone mode is not supported, deactivating the flag in"
+            "Starting Workflow UI in standalone mode is not supported, deactivating the flag in"
                 + " config");
         getLibConfig().activateWorkflowUi = false;
       }
@@ -460,6 +452,7 @@ public class TigerDirector {
     System.clearProperty("http.proxyPort");
     System.clearProperty("https.proxyPort");
 
+    ScenarioRunner.clearScenarios();
     TigerGlobalConfiguration.reset();
   }
 
@@ -536,13 +529,5 @@ public class TigerDirector {
           message,
           errorMessage);
     }
-  }
-
-  public static void registerRuntime(Runtime runtime) {
-    TigerDirector.runtime = runtime;
-  }
-
-  public static Optional<Runtime> loadRuntime() {
-    return Optional.ofNullable(TigerDirector.runtime);
   }
 }