+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/env/TestResult.java
@@ -24,5 +24,6 @@ public enum TestResult {
   EXECUTING,
   UNDEFINED,
   UNUSED,
-  AMBIGUOUS
+  AMBIGUOUS,
+  TEST_DISCOVERED
 }